backprojecting taa

+++ src/wrld/resources/TAAFramebuffer.cpp
@@ -21,7 +21,7 @@ namespace wrld::rsc {
         // position texture
         glGenTextures(1, &position_texture);
         glBindTexture(GL_TEXTURE_2D, position_texture);
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB16F, width, height, 0, GL_RGB, GL_FLOAT,
+        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB32F, width, height, 0, GL_RGB, GL_FLOAT,
                      nullptr);
 
         glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
